[refact] Rename according to google style
