+++ Shell/Common.h
@@ -10,6 +10,7 @@ enum ShellCommand {
 	ERASE_COMMAND,
 	ERASERANGE_COMMAND,
 	FLUSH_COMMAND,
+	SCRIPT_RUN_COMMAND
 };
 enum ShellScriptApiCommand {
 	WRITE_SCRIPT_COMMAND,
