+++ Shell/shell_exception.cpp
