+++ Shell/shell_script_write_command.cpp
@@ -1,8 +1,8 @@
 #pragma once
-#include "ShellScriptCommandInterface.h"
-#include "ShellUtil.cpp"
+#include "shell_script_command_interface.h"
+#include "shell_util.cpp"
 #include "ssddriver_interface.h"
-#include "ScriptParameterGenerator.h"
+#include "script_parameter_generator.h"
 #include <memory>
 
 using std::shared_ptr;
